requirements.txt
streamlit==1.28.0
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
xgboost==2.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
matplotlib==3.7.2
seaborn==0.12.2